Genmap code logic refactoring
